refactor(aerorsync): move the trait impls and maps to the adapter
Second of the three A3 commits. `src-tauri/src/aerorsync_adapter/` is new: it holds every AeroFTP type the module is no longer allowed to name, and the module stops naming them.

What moved, verbatim. `impl DeltaTransport for AerorsyncDeltaTransport` and `impl DeltaBatch for AerorsyncBatch` are now in `aerorsync_adapter/delta_transport.rs`; the maps onto the application error, statistics and capability types are in `aerorsync_adapter/errors.rs`, together with the nine tests that pin them (the seven fallback-policy pins and the two statistics pins) and the two tests this tranche added for the rendering and the retry parity. The envelope strings are untouched, and they are now built in exactly one place, which matters because `delta_sync_rsync` re-reads them to choose between a transparent classic fallback and a surfaced hard error.

What the module gained instead. `probe` returns the crate-owned `TransportProbe` and lets the caller decide what a failure means; `open_batch` returns an `AerorsyncBatch` and lets the caller decide what to do when the handshake fails; `endpoint` exposes the host and port the adapter names in its log lines. The two log lines themselves, the probe warning and the batch-degradation warning, moved with the code that owns the decision and print the same text.

Why the implementations move in this tranche and not the next. The trait methods are the only place that has to produce an `RsyncError`, so leaving them in the module would leave the maps there too, and the module cannot import its own adapter. Moving them costs nothing at the call sites: an implementation is resolved crate-wide, so `providers/sftp.rs`, `local_sync.rs`, `bin/aeroftp_cli.rs` and the integration tests keep compiling with zero lines changed.

The import budget drops to its A3 target: 5 keys and 12 lines, over two application modules. What remains is the `DeltaTransport` trait in the tests that drive a batch, the same trait in `live_tests.rs` and `local_transport.rs`, and `RsyncConfig` with `AuthMethod` in `from_rsync_config`, all of which belong to the next tranche. The test was seen red twice with the new constant, on a plain `use crate::rsync_over_ssh::RsyncStats;` and on `#[cfg(any())] use crate::aerorsync_adapter::errors::to_rsync_error;`, the second being the proof that the module may never reach for its own adapter. The adapter has the mirror-image guard: `aerorsync_adapter_reaches_only_the_module_public_surface` reads the adapter's own sources and refuses any module path outside a declared list.

Two notes from the previous tranche's review are closed here, both text only: the module's progress contract no longer spells the application alias in a comment, and the budget test no longer says the adapter is created by A2.

Counts: the module suite goes from 688 to 689 passed, 0 failed, 15 ignored, zero skipped oracle pins; the one new test is the adapter surface guard, which the `aerorsync` filter selects through the adapter's own path. `cargo fmt --all -- --check`, `cargo check --no-default-features --lib`, `cargo clippy --lib --features aerorsync --all-targets -- -D warnings`, `cargo check --bin aeroftp-cli --features aerorsync`, `cargo test --lib delta_sync_rsync`, `cargo test --lib rsync_over_ssh`, the offline hard-rejection string contract and `git diff --check` all pass.
